+++ docs/operators-and-hooks-ref.rst
@@ -353,7 +353,7 @@ These integrations allow you to perform various operations within the Amazon Web
      -
      - :mod:`airflow.providers.amazon.aws.hooks.emr`
      - :mod:`airflow.providers.amazon.aws.operators.emr_add_steps`,
-       :mod:`airflow.contrib.operators.emr_create_job_flow_operator`,
+       :mod:`airflow.providers.amazon.aws.operators.emr_create_job_flow`,
        :mod:`airflow.contrib.operators.emr_terminate_job_flow_operator`
      - :mod:`airflow.contrib.sensors.emr_base_sensor`,
        :mod:`airflow.contrib.sensors.emr_job_flow_sensor`,
